# perf: update lambdaworks and use Ruffini in deep composition poly