Updates lambdaworks dep to use Ruffini division in deep composition polynomials.
The update also includes a few other optimizations, namely the Merkle tree build one.

## Bench results (on M1)

```
CAIRO/fibonacci/10      time:   [59.898 ms 60.048 ms 60.284 ms]
                        change: [-53.124% -52.911% -52.675%] (p = 0.00 < 0.05)
CAIRO/fibonacci/30      time:   [132.44 ms 132.91 ms 133.17 ms]
                        change: [-66.094% -66.001% -65.920%] (p = 0.00 < 0.05)
```